pyproject.toml
[project]
name = "automate-discord-notification"
version = "0.1.0"
description = ""
authors = [{ name = "Ahda Akmalul Ilmi", email = "ahda043@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "google-cloud-firestore (>=2.21.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
